Fix some classes that were exposed outside their visibility scope

+++ modules/pj/src/main/java/edu/rit/pj/reduction/SharedInteger.java
@@ -241,7 +241,7 @@ public class SharedInteger
      * @return Current value.
      */
     public long longValue() {
-        return (long) get();
+        return get();
     }
 
     /**
